refactor CohortEditor and CohortEditorFilter
